+++ app/channels/organization_channel.rb
@@ -1,0 +1,9 @@
+class OrganizationChannel < ApplicationCable::Channel
+  def subscribed
+    stream_from "organization_#{params[:uuid]}"
+  end
+
+  def unsubscribed
+    # Any cleanup needed when channel is unsubscribed
+  end
+end