+++ app/controllers/api/ui/v1/course_events_controller.rb
@@ -36,6 +36,7 @@ class Api::Ui::V1::CourseEventsController < Api::Ui::BaseController
   end
 
   def show
+    @show_subscribed = true
     render :show
   end
 
@@ -51,6 +52,11 @@ class Api::Ui::V1::CourseEventsController < Api::Ui::BaseController
         attendee&.destroy
       end
     if status
+      # ActionCable.server.broadcast "room_#{@course_event.id}", attendees: @course_event.attendees.map { |a| { id: a.id, presence: a.presence, user_name: "#{a.user.firstname} #{a.user.lastname}" } }
+      @show_subscribed = false
+      # ActionCable.server.broadcast "room_#{@course_event.id}", course_event: JSON.parse(render_to_string(:show))
+      ActionCable.server.broadcast "organization_#{@organization.uuid}", course_event: JSON.parse(render_to_string(:show))
+      @show_subscribed = true
       render :show
     elsif attendee
       render json: { errors: attendee.errors }, status: :unprocessable_entity